add futures-util feature flag + impl Stream for Rows
separate statement to separate file

move rows to separate file

make futures-util default feature + use futures in example.rs

docs fix

reset statements when we are done with the query or we encounter an error

extra example using Streams mapping

more docs

clippy

fix doc test
