+++ bindings/rust/examples/example.rs
@@ -1,3 +1,6 @@
+#![cfg(feature = "futures")]
+
+use futures_util::stream::TryStreamExt;
 use turso::Builder;
 
 #[tokio::main]
@@ -6,8 +9,15 @@ async fn main() {
 
     let conn = db.connect().unwrap();
 
-    conn.query("select 1; select 1;", ()).await.unwrap();
+    // `query` and other methods, only parse the first query given.
+    let mut rows = conn.query("select 1; select 1;", ()).await.unwrap();
+    // Iterate over the rows with the Stream iterator syntax
+    while let Some(row) = rows.try_next().await.unwrap() {
+        let val = row.get_value(0).unwrap();
+        println!("{:?}", val);
+    }
 
+    // Contrary to `prepare` and `query`, `execute` is not lazy and will execute the query to completion
     conn.execute("CREATE TABLE IF NOT EXISTS users (email TEXT)", ())
         .await
         .unwrap();
@@ -32,9 +42,20 @@ async fn main() {
 
     let mut rows = stmt.query(["foo@example.com"]).await.unwrap();
 
-    let row = rows.next().await.unwrap().unwrap();
+    while let Some(row) = rows.try_next().await.unwrap() {
+        let value = row.get_value(0).unwrap();
+        println!("Row: {:?}", value);
+    }
 
-    let value = row.get_value(0).unwrap();
+    let rows = stmt.query(["foo@example.com"]).await.unwrap();
 
-    println!("Row: {:?}", value);
+    // As `Rows` implement streams you can easily map over it and apply other transformations you see fit
+    println!("Using Stream map");
+    rows.map_ok(|row| row.get_value(0).unwrap())
+        .try_for_each(|val| {
+            println!("Row: {:?}", val.as_text().unwrap());
+            futures_util::future::ready(Ok(()))
+        })
+        .await
+        .unwrap();
 }
